A new failing test for finding elements by some property has revealed that we cannot have drivers depend on concrete UIView instances, as they may not be available at the time the driver is created.
We will need to work towards some kind of asynchronous probing for views
but we can get part of the way there by having drivers depend on an object
that implements the AKViewSelector protocol.

A couple of implementations are provided so far; one that simply wraps
an actual known UIView that we know exists at runtime, one that finds
a subview using its tag and one that finds the main window.